frontend/requirements.txt
# Frameworks for UI
gradio==3.40.0
flask==2.3.3

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Kafka for live updates
kafka-python==2.1.0

# Cloudinary for image upload/display
cloudinary==1.32.0

# Optional: JSON handling (רוב הפעמים כבר מובנה)
ujson==5.8.0

# Optional: Data structures for Models
pydantic==2.5.1
